+++ src/utils/Utils.js
@@ -2,6 +2,22 @@
  * @flow
  */
 
+import entity from '../../testData/entity.json';
+import property from '../../testData/property.json';
+
+
+export function getMoreDetails(association :object) {
+  const details = {};
+
+  details.Source = association.src.map(number => entity.find(el => el.id === number));
+  details.Destination = association.dst.map(number => entity.find(el => el.id === number));
+
+  details.Key = association.entityType.key.map(number => property.find(el => el.id === number));
+  details.Properties = association.entityType.properties.map(number => property.find(el => el.id === number));
+
+  return details;
+}
+
 export function randomId() :string {
 
   // https://stackoverflow.com/questions/105034/create-guid-uuid-in-javascript
